+++ src/reagent/ratom.cljs
@@ -34,10 +34,20 @@
                  false))))))
 
 (defn- in-context [obj f]
+  "When f is executed, if (f) derefs any ratoms, they are then added to 'obj.captured'(*ratom-context*).
+
+   See function notify-deref-watcher! to know how *ratom-context* is updated"
   (binding [*ratom-context* obj]
     (f)))
 
-(defn- deref-capture [f r]
+(defn- deref-capture
+  "Returns `(in-context f r)`.  Calls `_update-watching` on r with any
+   `deref`ed atoms captured during `in-context`, if any differ from the
+   `watching` field of r.  Clears the `dirty?` flag on r.
+
+   Inside '_update-watching' along with adding the ratoms in 'r.watching' of reaction,
+   the reaction is also added to the list of watches on each ratoms f derefs."
+  [f r]
   (set! (.-captured r) nil)
   (when (dev?)
     (set! (.-ratomGeneration r) (set! generation (inc generation))))
@@ -49,7 +59,11 @@
       (._update-watching r c))
     res))
 
-(defn- notify-deref-watcher! [derefed]
+(defn- notify-deref-watcher!
+  "Add `derefed` to the `captured` field of `*ratom-context*`.
+
+  See also `in-context`"
+  [derefed]
   (when-some [r *ratom-context*]
     (let [c (.-captured r)]
       (if (nil? c)
@@ -336,7 +350,15 @@
 (defn- handle-reaction-change [this sender old new]
   (._handle-change this sender old new))
 
-
+;; Fields of a Reaction javascript object
+;; - auto_run
+;; - captured
+;; - caught
+;; - f
+;; - ratomGeneration
+;; - state
+;; - watches
+;; - watching
 (deftype Reaction [f ^:mutable state ^:mutable ^boolean dirty? ^boolean nocache?
                    ^:mutable watching ^:mutable watches ^:mutable auto-run
                    ^:mutable caught]
@@ -498,7 +520,16 @@
 
 (def ^:private temp-reaction (make-reaction nil))
 
-(defn run-in-reaction [f obj key run opts]
+
+(defn run-in-reaction
+  "Evaluates `f` and returns the result.  If `f` calls `deref` on any ratoms,
+   creates a new Reaction that watches those atoms and calls `run` whenever
+   any of those watched ratoms change.  Also, the new reaction is added to
+   list of 'watches' of each of the ratoms. The `run` parameter is a function
+   that should expect one argument.  It is passed `obj` when run.  The `opts`
+   are any options accepted by a Reaction and will be set on the newly created
+   Reaction. Sets the newly created Reaction to the `key` on `obj`."
+  [f obj key run opts]
   (let [r temp-reaction
         res (deref-capture f r)]
     (when-not (nil? (.-watching r))